Wrote Feature that allows us to switch the networkpath (#39)
* Wrote Feature that allows us to switch the networkpath
We can open a new console and model on defined pre-existing network file. The python side does extract the path and opens a new console. Only issue is that resulting calls to run a second time. The pipe is open though.
Also moved the opening of the console and model into its own function called loadModel making it more flexible and robust

* Re-factored the code based on the Github feedback.
Now the switchModel() does run and close successfully as intended.

* Re-factored the json parser
Passing parameters more flexibly.
Still need to clean up the codebase a bit more

* Removed dead code on line 515 in importnetwork.py

* removed the aimsuncontroller debuggerattached commented code as we don't need it anymore
changed SwitchNetworkPath to SignalSwitchNetworkPath to follow appropaite naming convention
added comments to the TestConfiguration class

* added space between methods in aimsuncontroller.cs
testconfiguration helper class comments are triple slashed for better documentation
inlined the json creation data
removed second unittest as it was redundant and not needed

* fixed the grammar typos and removed the launchaimsun input variable as it is not needed anymore