Refactored macroAssignment.py (#46)
* Refactored macroAssignment.py
Refactored codebae to work from unittest bridge and terminal. Added a new unittest module to test macroAssignment.
Note that the split of road and travel will be done in the next PR cycle upon further clarification.

* Fixed runFromConsole() function. Removed duplicate code and now calls _execute() function

* Refactored the script into several module suctions.
Broke the code based on the comments.
Note that the save_network() function has several input parameters. It may be required to pass in a class to keep it more flexible.
ToDo: split the transit and road assignment but not sure how to do that yet.

* Broke the experiment from the create_scenario
created new functions for create_scenario and pt_scenario

* Resolved the errors  in the module
removed the duplicate return statement, renamed function to be more descriptive
created a new function create_schedule_demand_item and moved associated code out of _execute

* Moved the xtmf_parameters hard coded values out of the code and made them passable by the bridge.
Line 30 name of transit comes from the xtmf_parameters

* renamed the dictionary of matrix to autoDemand

* formatted the document so it has proper spacing and following correct guidelines